requirements.txt
flask==2.3.3
python-dotenv==1.0.0
supabase==1.2.0
gunicorn==21.2.0
httpx==0.24.1
pandas==2.2.3
openpyxl==3.1.2
pymongo==4.6.0
flask-pymongo==2.3.0
flask-login==0.6.2
bcrypt==4.0.1